requirements.txt
discord.py==2.4.0
aiohttp==3.10.11
beautifulsoup4==4.12.3
aiosqlite==0.20.0
lxml==5.3.0
